--- a/package/MDAnalysis/analysis/rms.py
+++ b/package/MDAnalysis/analysis/rms.py
@@ -147,8 +147,8 @@ import MDAnalysis.lib.qcprot as qcp
 from MDAnalysis.analysis.base import AnalysisBase
 from MDAnalysis.exceptions import SelectionError, NoDataError
 from MDAnalysis.lib.log import ProgressMeter
-from MDAnalysis.lib.util import asiterable, iterable, get_weights, deprecate
-
+from MDAnalysis.lib.util import (asiterable, iterable, get_weights, deprecate,
+                                 coords_add_vec)
 
 logger = logging.getLogger('MDAnalysis.analysis.rmsd')
 
@@ -621,11 +621,11 @@ class RMSD(AnalysisBase):
             # Transform each atom in the trajectory (use inplace ops to
             # avoid copying arrays) (Marginally (~3%) faster than
             # "ts.positions[:] = (ts.positions - x_com) * R + ref_com".)
-            self._ts.positions[:] -= mobile_com
+            coords_add_vec(self._ts.positions, -mobile_com)
 
             # R acts to the left & is broadcasted N times.
             self._ts.positions[:] = np.dot(self._ts.positions, self._R)
-            self._ts.positions += self._ref_com
+            coords_add_vec(self._ts.positions, self._ref_com)
 
             # 2) calculate secondary RMSDs (without any further
             #    superposition)