… package/MDAnalysis/core/topologyattrs.py
… package/MDAnalysis/core/universe.py
… package/MDAnalysis/lib/_cutil.pyx
… package/MDAnalysis/transformations/rotate.py
… package/MDAnalysis/transformations/translate.py
… testsuite/MDAnalysisTests/core/test_atomgroup.py
… testsuite/MDAnalysisTests/core/test_fragments.py
… testsuite/MDAnalysisTests/core/test_groups.py
… testsuite/MDAnalysisTests/core/test_wrap.py
… testsuite/MDAnalysisTests/lib/test_cutil.py
… testsuite/MDAnalysisTests/lib/test_util.py