Fixes #2210 

Changes made in this Pull Request:
 - Changed `coordinates.base.Timestep.order` to 'C'
 - Added optional keywords `return_counts`, `return_masks`, and `assume_unsorted` to `lib.util.unique_int_1d()` and improved its performance
 - Added `lib.util.iscontiguous_int_1d()`
 - Added `lib.util.indices_to_slice_1d()`
 - Added `*Group.iscontiguous` property
 - Added `*Group._ix_or_slice` property
 - Added `AtomGroup._positions_view_or_copy` property returning a view for sliceable AtomGroups (otherwise a copy)
 - Added `AtomGroup._positions_c_view_or_copy` property returning a C-contiguous view for contiguous AtomGroups (otherwise a copy)
 - Added `lib.util.unique_masks_int_1d()`
 - Added `lib.util.argwhere_int_1d()` (will be used in a future PR for faster unwrapping)
 - Added `lib.util.check_compound()`
 - Added `AtomGroup.compound_indices()`
 - Added `lib.util.coords_add_vector()`
 - Added `lib.util._coords_add_vectors()`
 - Replaced `Universe._fragdict` by `Universe._fraginfo` property (accelerates `AtomGroup.fragments` and `AtomGroup.fragindices`)
 - Improved performance of `*Group.center()`, especially for per-compound computations and contiguous AtomGroups
 - Improved performance of `*Group.wrap()`, especially for per-compound computations and  contiguous AtomGroups (up to a factor of **39.3**)

PR Checklist
------------
 - [x] Tests?
 - [x] Docs?
 - [x] CHANGELOG updated?
 - [x] Issue raised/referenced?