@zemanj 2019-02-28
In some tests, Travis/conda/whatever seems to screw up:
`unable to execute 'x86_64-conda_cos6-linux-gnu-gcc': No such file or directory`