@zemanj 2019-03-26
## Performance Report ##

### Test Code ###
```Python
# -*- coding: utf-8 -*-
import numpy as np
import MDAnalysis as mda
from timeit import timeit


########## HELPER FUNCTIONS ##########

def run(func, nframes):
    for ts in u.trajectory[:nframes]:
        func()

def trajectory_read(nframes):
    for ts in u.trajectory[:nframes]:
        pass

def time_group_wrap(group, compound, center, inplace, nruns, nframes_per_run,
                    t_read):
    func = lambda: group.wrap(compound=compound, center=center, inplace=inplace)
    t_wrap = timeit(lambda: run(func, nframes_per_run), number=nruns)
    t_wrap *= 1.0e6 / (nruns * nframes_per_run)
    t = t_wrap - t_read
    print("| {:9s} | {:6s} | {:7s} | {:10.1f} |"
          "".format(compound, center, str(inplace), t))

def run_measurements(groups, compounds, centers, nruns, nframes_per_run):
    t_read = timeit(lambda: trajectory_read(nframes_per_run), number=nruns)
    t_read *= 1.0e6 / (nruns * nframes_per_run)
    for cont, group in groups:
        print("\n{} {}\n".format(cont, repr(group).strip("<>")))
        print("| compound  | center | inplace | µs / frame |")
        print("| :---      | :---:  | :---    |       ---: |")
        for compound in compounds:
            if compound == 'atoms':
                _centers = ['n/a']
            else:
                _centers = centers
            for center in _centers:
                for inplace in [False, True]:
                    time_group_wrap(group, compound, center, inplace,
                                    nruns, nframes_per_run, t_read)


########## SETUP ##########

# Load Universe:
u = mda.Universe("topol.tpr", "traj_comp.xtc")
# populate u._fraginfo:
_ = u.atoms.fragments

# Set up groups, compounds, and centers:
np.random.seed(1337)
rnd_aix = u.atoms.ix.copy()
np.random.shuffle(rnd_aix)
groups = [("contiguous", u.atoms),
          ("non-contiguous", u.atoms[::-1]),
          ("scrambled", u.atoms[rnd_aix])]
compounds = ['atoms', 'group', 'segments', 'residues', 'molecules', 'fragments']
centers = ['com', 'cog']

# measurement parameters:
nruns = 1000
nframes_per_run = 100

print("{} runs, {} frames / run".format(nruns, nframes_per_run))


########## MEASUREMENTS ##########

run_measurements(groups, compounds, centers, nruns, nframes_per_run)
```

### Runtime Comparison ###
Test CPU: AMD Ryzen Threadripper 1920X 12-Core Processor @ 3.5 GHz
Test data: GROMACS xtc trajectory from a simulation of Neat BMIm BF4, 500 ion pairs (15000 atoms) in a fully periodic cubic unit cell (NpT ensemble)
Test configuration: 1000 runs, 100 frames / run

Note that the runtimes presented below are execution times of `AtomGroup.wrap()` only, i.e., the time needed for trajectory reading time has been subtracted.

#### contiguous AtomGroup with 15000 atoms (`u.atoms`) ####

| compound    | center | inplace | µs / frame (new) | µs / frame (old) | speedup factor (old/new) |
| :---        | :---:  | :---    |             ---: |             ---: |                     ---: |
| atoms       | n/a    | False   |            199.6 |            211.3 |                     1.06 |
| atoms       | n/a    | True    |             49.8 |            394.7 |                     7.93 |
| group       | com    | False   |            238.2 |            722.4 |                     3.03 |
| group       | com    | True    |             85.3 |            905.6 |                    10.62 |
| group       | cog    | False   |            206.4 |            599.8 |                     2.91 |
| group       | cog    | True    |             56.0 |            782.7 |                    13.98 |
| segments    | com    | False   |            489.3 |           2011.1 |                     4.11 |
| segments    | com    | True    |            339.9 |           2194.2 |                     6.46 |
| segments    | cog    | False   |            457.6 |           1858.4 |                     4.06 |
| segments    | cog    | True    |            299.7 |           2041.2 |                     6.81 |
| residues    | com    | False   |           1198.2 |          33676.5 |                    28.11 |
| residues    | com    | True    |           1035.7 |          33874.6 |                    32.71 |
| residues    | cog    | False   |           1147.2 |          33407.9 |                    29.12 |
| residues    | cog    | True    |            994.9 |          33745.0 |                    33.92 |
| molecules   | com    | False   |            843.2 |          18311.0 |                    21.72 |
| molecules   | com    | True    |            690.1 |          18520.9 |                    26.84 |
| molecules   | cog    | False   |            803.5 |          18148.0 |                    22.59 |
| molecules   | cog    | True    |            650.1 |          18377.9 |                    28.27 |
| fragments   | com    | False   |            829.2 |          25149.5 |                    30.33 |
| fragments   | com    | True    |            674.4 |          25387.2 |                    37.64 |
| fragments   | cog    | False   |            788.2 |          24823.8 |                    31.49 |
| fragments   | cog    | True    |            635.7 |          25004.9 |                    39.33 |
| **average** | **com/cog** | **False** |        **654.6** |      **14447.2** |                **22.07** |
| **average** | **com/cog** | **True** |        **501.1** |      **14657.2** |                **29.25** |
| **average** | **com/cog** | **False/True** |        **577.8** |      **14552.2** |                **25.18** |


#### non-contiguous AtomGroup with 15000 atoms (`u.atoms[::-1]`) ####

| compound    | center | inplace | µs / frame (new) | µs / frame (old) | speedup factor (old/new) |
| :---        | :---:  | :---    |             ---: |             ---: |                     ---: |
| atoms       | n/a    | False   |            197.9 |            213.0 |                     1.08 |
| atoms       | n/a    | True    |            176.8 |            415.9 |                     2.35 |
| group       | com    | False   |            233.3 |            724.2 |                     3.10 |
| group       | com    | True    |            140.6 |            906.3 |                     6.45 |
| group       | cog    | False   |            205.5 |            606.5 |                     2.95 |
| group       | cog    | True    |            112.7 |            793.2 |                     7.04 |
| segments    | com    | False   |            479.4 |           2026.0 |                     4.23 |
| segments    | com    | True    |            386.6 |           2206.9 |                     5.71 |
| segments    | cog    | False   |            440.4 |           1872.4 |                     4.25 |
| segments    | cog    | True    |            348.1 |           2062.9 |                     5.93 |
| residues    | com    | False   |           1183.0 |          33750.9 |                    28.53 |
| residues    | com    | True    |           1079.9 |          33996.4 |                    31.48 |
| residues    | cog    | False   |           1136.6 |          33460.7 |                    29.44 |
| residues    | cog    | True    |           1040.0 |          33677.7 |                    32.38 |
| molecules   | com    | False   |            854.0 |          18210.1 |                    21.32 |
| molecules   | com    | True    |            766.6 |          18426.2 |                    24.04 |
| molecules   | cog    | False   |            816.3 |          18034.4 |                    22.09 |
| molecules   | cog    | True    |            726.9 |          18206.0 |                    25.05 |
| fragments   | com    | False   |            840.0 |          25127.3 |                    29.91 |
| fragments   | com    | True    |            753.1 |          25364.3 |                    33.68 |
| fragments   | cog    | False   |            798.9 |          24892.8 |                    31.16 |
| fragments   | cog    | True    |            709.5 |          25024.5 |                    35.27 |
| **average** | **com/cog** | **False** |        **653.2** |      **14447.1** |                **22.12** |
| **average** | **com/cog** | **True** |        **567.3** |      **14643.7** |                **25.81** |
| **average** | **com/cog** | **False/True** |        **610.3** |      **14545.4** |                **23.83** |

#### scrambled AtomGroup with 15000 atoms (`u.atoms[shuffled_indices]`) ####

| compound    | center | inplace | µs / frame (new) | µs / frame (old) | speedup factor (old/new) |
| :---        | :---:  | :---    |             ---: |             ---: |                     ---: |
| atoms       | n/a    | False   |            201.7 |            213.2 |                     1.06 |
| atoms       | n/a    | True    |            378.4 |            396.1 |                     1.05 |
| group       | com    | False   |            238.7 |            723.0 |                     3.03 |
| group       | com    | True    |            411.8 |            904.5 |                     2.20 |
| group       | cog    | False   |            217.0 |            599.8 |                     2.76 |
| group       | cog    | True    |            383.6 |            784.2 |                     2.04 |
| segments    | com    | False   |            538.4 |           2015.5 |                     3.74 |
| segments    | com    | True    |            705.2 |           2201.4 |                     3.12 |
| segments    | cog    | False   |            489.3 |           1863.1 |                     3.81 |
| segments    | cog    | True    |            670.7 |           2047.4 |                     3.05 |
| residues    | com    | False   |           1771.1 |          33739.9 |                    19.05 |
| residues    | com    | True    |           1940.2 |          33851.5 |                    17.45 |
| residues    | cog    | False   |           1719.7 |          33423.1 |                    19.44 |
| residues    | cog    | True    |           1892.0 |          34512.8 |                    18.24 |
| molecules   | com    | False   |           1377.8 |          18672.9 |                    13.55 |
| molecules   | com    | True    |           1546.2 |          18348.2 |                    11.87 |
| molecules   | cog    | False   |           1323.1 |          17975.6 |                    13.59 |
| molecules   | cog    | True    |           1494.7 |          18158.8 |                    12.15 |
| fragments   | com    | False   |           1359.3 |          25020.9 |                    18.41 |
| fragments   | com    | True    |           1528.4 |          25237.3 |                    16.51 |
| fragments   | cog    | False   |           1319.2 |          24755.0 |                    18.77 |
| fragments   | cog    | True    |           1485.6 |          24898.2 |                    16.76 |
| **average** | **com/cog** | **False** |        **959.6** |      **14454.7** |                **15.06** |
| **average** | **com/cog** | **True** |        **1130.6** |      **14667.3** |                **12.97** |
| **average** | **com/cog** | **False/True** |       **1045.1** |      **14561.0** |                **13.93** |